+++ backend/database/python/genre/helpers.py
@@ -1,0 +1,10 @@
+def format_genres(genre_data):
+    genres = []
+    for row in genre_data:
+        genre = {
+            "id": row[0],
+            "genre": row[1],
+        }
+        genres.append(genre)
+
+    return genres
